Fixed texture mapping when building faces and added normal coordinates.

+++ src/World/Block.hpp
@@ -28,7 +28,7 @@ public:
     void setType(BlockType type);
 
 public:
-    static Vertex buildFace(GLfloat v0[3], GLfloat v1[3], GLfloat v2[3], GLfloat v3[3], int width, int height,
+    static Vertex buildFace(GLfloat v0[3], GLfloat v1[3], GLfloat v2[3], GLfloat v3[3], GLfloat width, GLfloat height,
             face side, glm::ivec3 chunkPos, unsigned int vertexCount, BlockType type);
 };
